HW9-Hibernate - fixed the typos in TicketDAO (was TicketsDAO) and in UserDAO (was UsersDAO)
