+++ src/main/java/com/kuzmins/dao/TicketDAO.java
@@ -1,42 +1,58 @@
 package com.kuzmins.dao;
 
+import com.kuzmins.model.Ticket;
 import com.kuzmins.model.User;
 import org.hibernate.Session;
 import org.hibernate.Transaction;
 
+import java.util.ArrayList;
+import java.util.List;
 import java.util.UUID;
 
-public class UsersDAO {
+public class TicketDAO {
 
-    public void saveUser(User user) {
+    public void saveTicket(Ticket ticket) {
         Session session = SessionFactoryProvider.getSessionFactory().openSession();
         Transaction transaction = session.beginTransaction();
-        session.save(user);
+        session.save(ticket);
         transaction.commit();
         session.close();
     }
 
-    public User getUserById(UUID id) {
+    public Ticket getTicketById(UUID id) {
+        Session session = SessionFactoryProvider.getSessionFactory().openSession();
+        Transaction transaction = session.beginTransaction();
+        Ticket ticket = session.get(Ticket.class, id);
+        transaction.commit();
+        session.close();
+        return ticket;
+    }
+
+    public List<Ticket> getTicketsByUserId(UUID id) {
+        List<Ticket> tickets = new ArrayList<>();
         Session session = SessionFactoryProvider.getSessionFactory().openSession();
         Transaction transaction = session.beginTransaction();
         User user = session.get(User.class, id);
+        if (user.getTickets() != null) {
+            tickets = user.getTickets();
+        }
         transaction.commit();
         session.close();
-        return user;
+        return tickets;
     }
 
-    public void updateUser(User user) {
+    public void updateTicket(Ticket ticket) {
         Session session = SessionFactoryProvider.getSessionFactory().openSession();
         Transaction transaction = session.beginTransaction();
-        session.update(user);
+        session.update(ticket);
         transaction.commit();
         session.close();
     }
 
-    public void deleteUser(User user) {
+    public void deleteTicket(Ticket ticket) {
         Session session = SessionFactoryProvider.getSessionFactory().openSession();
         Transaction transaction = session.beginTransaction();
-        session.delete(user);
+        session.delete(ticket);
         transaction.commit();
         session.close();
     }
